Proportional Hazards Mixture Cure Model (MCM)

Estimates a PH-MCM with EM algorithm. Returns coefficients from survival (beta) and cure (gamma) models. For reference to notation see http://post.queensu.ca/~pengp/papers/Peng2014_4.pdf.

Usage

CUREcoxph(survform,cureform,So,X,bootstrap=NULL,verbose=TRUE,subset=NULL,tol=1e-3,max.iter=50);

Arguments

survform

A character or formula that is passed to survival model

cureform

A character or formula that is passed to cure model

So

A Surv() object

data

A data.frame with columns that contain the survform + cureform variables

bootstrap

Number of bootstrap iterations to conduct; if NULL, does not execute

verbose

Prints updates during function execution

subset

Subsets dat to this index

tol

L2 parameter tolerance

max.iter

Maximum number of iterations during EM algorithm

Value

Returns a list with parameters from the survival (beta) and cure (gamma) models, along with 95 percent CI if the bootstrap argument is used.

Author

Erik Drysdale

Examples

# # Simple example 
# library(survival)
# library(BoutrosLab.statistics.survival)
# df <- survival::cancer
# So <- with(df,Surv(time,status==2))
# # Normal Cox model
# mdl.coxph <- coxph(So ~ age + factor(sex),data=df)
# # Cure model
# survform <- '~age+factor(sex)'
# cureform <- '~age+factor(sex)'
# mdl.cure <- CUREcoxph(survform,cureform,So,df,max.iter = 250)
# coef.compare <- data.frame(coxph=c(NA,coef(mdl.coxph)),CUREcox=c(NA,mdl.cure$beta),CUREprob=mdl.cure$gamma)
# rownames(coef.compare) <- c('intercept','age','sex')
# print(round(coef.compare,3))
# 
# # Run Bootstrap
# mdl.bs <- CUREcoxph(survform,cureform,So,df,bootstrap=99,verbose=F,max.iter=250)
# # Show 95 percent CI for PH coefficients (age is insignificant)
# print(mdl.bs$beta)