Fit a Cox proportional hazards model

Fit a Cox model (possibly with some linear adjustments) and return key statistics about the fit.

Usage

fit.coxmodel(groups, survobj, other.data = NULL, stratification.factor, 
    stratification.value, rounding = 3, return.cox.model);

Arguments

groups

Grouping of patients (passed directly to coxph, so factors & continuous variables are okay)

survobj

An object of class Surv (from the survival package) -- patient ordering needs to be identical as for groups

other.data

A data-frame (or matrix?) of variables to be controlled in the Cox model. If null, no adjustment is done. No interactions are fit.

stratification.factor

A vector (or factor) containing a categorical variable on which the cox model will be stratified. If there are too many levels, it is possible that the model will not converge. Use return.cox.model = TRUE to verify if model looks reasonable.

stratification.value

A pre-determined value used to split the stratification factor. Must be within the range of the variable. (i.e. strata(stratification.factor > stratification.value)).

rounding

Specifies the number of digits of precision to be returned. Defaults to 3.

return.cox.model

If TRUE, returns coxph object. If FALSE, returns summary statistics only. Defaults to FALSE

Value

A vector containing: HR, lower 95% CI of HR, upper 95% CI of HR, P-value (for groups), number of samples (total with group assignments, although some may not be included in fit for other reasons so this is an upper-limit).

Author

Paul C. Boutros

Examples

set.seed(151)
survtime <- sample(seq(0.1,10,0.1), 100, replace = TRUE);
survstat <- sample(c(0,1), 100, replace = TRUE);
survobj <- Surv(survtime, survstat);
groups <- sample(c('A','B'), 100, replace = TRUE);
gender <- factor(sample(c('Male','Female'), 100, replace = TRUE), levels=c("Male","Female"));
age <- factor(sample(c('Young','MiddleAged','Old'), 100, replace = TRUE), 
            levels = c('Young', 'MiddleAged', 'Old'));
numbered.age <- sample(seq(40,90), 100, replace = TRUE);

fit.coxmodel(
  groups = as.factor(groups),
  survobj = survobj
  );
#> [1]   0.9820000   0.5680000   1.6990000   0.9489288 100.0000000

fit.coxmodel(
  groups = as.factor(groups),
  survobj = survobj,
  other.data = data.frame(age, gender),
  return.cox.model = TRUE  
  )
#> Call:
#> coxph(formula = survobj ~ ., data = variables)
#> 
#>                    coef exp(coef)  se(coef)      z     p
#> groupsB       -0.006738  0.993284  0.283436 -0.024 0.981
#> ageMiddleAged  0.309967  1.363381  0.334560  0.926 0.354
#> ageOld         0.085883  1.089679  0.357212  0.240 0.810
#> genderFemale  -0.011759  0.988310  0.279433 -0.042 0.966
#> 
#> Likelihood ratio test=0.93  on 4 df, p=0.9205
#> n= 100, number of events= 53 

fit.coxmodel(
  groups = as.factor(groups),
  survobj = survobj,
  other.data = data.frame(age),
  stratification.factor = gender,
  return.cox.model = TRUE
  )
#> Call:
#> coxph(formula = survobj ~ . + strata(stratification.factor), 
#>     data = variables)
#> 
#>                  coef exp(coef) se(coef)     z     p
#> groupsB       0.01045   1.01050  0.28531 0.037 0.971
#> ageMiddleAged 0.34932   1.41811  0.33385 1.046 0.295
#> ageOld        0.06706   1.06936  0.36314 0.185 0.853
#> 
#> Likelihood ratio test=1.23  on 3 df, p=0.7457
#> n= 100, number of events= 53 

fit.coxmodel(
  groups = as.factor(groups),
  survobj = survobj,
  other.data = data.frame(gender),
  stratification.factor = numbered.age,
  stratification.value = 60,
  return.cox.model = TRUE
  )
#> Call:
#> coxph(formula = survobj ~ . + strata(stratification.factor > 
#>     stratification.value), data = variables)
#> 
#>                  coef exp(coef) se(coef)      z     p
#> groupsB       0.01344   1.01353  0.28130  0.048 0.962
#> genderFemale -0.06420   0.93782  0.28476 -0.225 0.822
#> 
#> Likelihood ratio test=0.05  on 2 df, p=0.9738
#> n= 100, number of events= 53