multi.point.HR.table

A function to estimate HR at different time points (time windows) when Cox PH assumption fails.

Usage

multi.point.HR.table(all.groups = NULL, all.survtime = NULL, all.survstat = NULL,
     truncation.thresholds = c(5, 10), covariates = NULL)

Arguments

all.groups

Grouping of patients (A vector containing survival groups

all.survtime

A vector containing survival times

all.survstat

A vector containing survival status

truncation.thresholds

A vector of end points of time windows for which HR is estimated

covariates

A data frame with covariates to be included in the cox model

Value

A matrix of HR, 95L CI, 95U CI, Wald P, n (samples) and cox.zph P for the full time frame, and for each time window established using truncation.thresholds

Examples

# set up survival data
set.seed(100);

survtime.ph <- c(
  sample(1:100, 100, replace = TRUE),
  sample(1:60, 50, replace = TRUE),
  sample(40:100, 50, replace = TRUE)
  );

survstat.ph <- c(
  sample( c(0,1), 100, replace = TRUE),
  sample( c(0,0,0,0,1), 50, replace = TRUE),
  sample( c(0,1,1,1,1), 50, replace = TRUE)
  );
groups.ph <- factor(c(rep('A', 100), rep('B', 100)), levels = c('A','B'));

# try example without covariates, specifying cut points
multi.point.HR.table(
  all.groups = groups.ph,
  all.survtime = survtime.ph,
  all.survstat = survstat.ph,
  truncation.thresholds = c(60)
  );
#>          n    wald.p global.ph.p  group_HR group_CI95l group_CI95h group_cox.p
#> Time60 200 0.5658418  0.07246465 0.8728447   0.5486906    1.388502   0.5658418
#>        group_ph.p
#> Time60 0.07246465

# now try example with covariates
gender.ph <- factor(sample(c("Male","Female"), 200, replace = TRUE), levels = c("Male","Female"));
score.ph <- rgamma(n=200, shape = 20, rate=1); # this is just an arbitrary "continuous" variable
multi.point.HR.table(
  all.groups = groups.ph,
  all.survtime = survtime.ph/12, 
  all.survstat = survstat.ph,
  covariates = data.frame(gender = gender.ph, score = score.ph)
  );
#> [1] "Warning: max survtime was less than max cut point. New cut points are 4.16666666666667"
#> [2] "Warning: max survtime was less than max cut point. New cut points are 8.33333333333333"
#>                        n    wald.p global.ph.p groupsB_HR groupsB_CI95l
#> Time4.16666666666667 200 0.3688150   0.4792561  0.7651198     0.4479911
#> Time8.33333333333333  90 0.3259063   0.1219341  1.6357638     0.9398838
#>                      groupsB_CI95h groupsB_cox.p groupsB_ph.p genderFemale_HR
#> Time4.16666666666667      1.306741    0.32692639    0.2072108        0.760277
#> Time8.33333333333333      2.846866    0.08174245    0.8752291        1.176518
#>                      genderFemale_CI95l genderFemale_CI95h genderFemale_cox.p
#> Time4.16666666666667          0.4468397           1.293576          0.3121581
#> Time8.33333333333333          0.6781721           2.041067          0.5630408
#>                      genderFemale_ph.p  score_HR score_CI95l score_CI95h
#> Time4.16666666666667         0.7255918 1.0239700   0.9767534    1.073469
#> Time8.33333333333333         0.3170440 0.9917997   0.9377078    1.049012
#>                      score_cox.p score_ph.p
#> Time4.16666666666667   0.3253944 0.32507706
#> Time8.33333333333333   0.7735293 0.02574333